removing pointless classes
