+++ autoarray/__init__.py
@@ -62,7 +62,6 @@ from .structures.grids.two_d.grid_2d_irregular import Grid2DIrregular
 from .structures.grids.two_d.grid_2d_irregular import Grid2DIrregularUniform
 from .structures.grids.two_d.grid_2d_pixelization import Grid2DRectangular
 from .structures.grids.two_d.grid_2d_pixelization import Grid2DVoronoi
-from .structures.grids.two_d.grid_2d_pixelization import Grid2DVoronoiNN
 from .structures.grids.two_d.grid_2d_pixelization import Grid2DDelaunay
 from .structures.vectors.uniform import VectorYX2D
 from .structures.vectors.irregular import VectorYX2DIrregular
