+++ autoarray/inversion/pixelizations/__init__.py
@@ -1,7 +1,7 @@
 from .rectangular import Rectangular
 from .voronoi import VoronoiMagnification
 from .voronoi import VoronoiBrightnessImage
-from .voronoi_nn import VoronoiNNMagnification
-from .voronoi_nn import VoronoiNNBrightnessImage
+from .voronoi import VoronoiNNMagnification
+from .voronoi import VoronoiNNBrightnessImage
 from .delaunay import DelaunayMagnification
 from .delaunay import DelaunayBrightnessImage
