+++ autoarray/structures/grids/two_d/grid_2d_pixelization.py
@@ -234,27 +234,7 @@ class Grid2DRectangular(AbstractStructure2D):
         )
 
 
-class Grid2DVoronoi(AbstractStructure2D):
-    """
-    Returns the geometry of the Voronoi pixelization, by alligning it with the outer-most coordinates on a \
-    grid plus a small buffer.
-
-    Parameters
-    -----------
-    grid
-        The (y,x) grid of coordinates which determine the Voronoi pixelization's
-    pixelization_grid
-        The (y,x) centre of every Voronoi pixel in scaleds.
-    origin
-        The scaled origin of the Voronoi pixelization's coordinate system.
-    pixel_neighbors
-        An array of length (voronoi_pixels) which provides the index of all neighbors of every pixel in \
-        the Voronoi grid (entries of -1 correspond to no neighbor).
-    pixel_neighbors.sizes
-        An array of length (voronoi_pixels) which gives the number of neighbors of every pixel in the \
-        Voronoi grid.
-    """
-
+class AbstractGrid2DIrregular(AbstractStructure2D):
     def __new__(
         cls,
         grid: Union[np.ndarray, List],
@@ -301,6 +281,28 @@ class Grid2DVoronoi(AbstractStructure2D):
                 obj.nearest_pixelization_index_for_slim_index
             )
 
+
+class Grid2DVoronoi(AbstractGrid2DIrregular):
+    """
+    Returns the geometry of the Voronoi pixelization, by alligning it with the outer-most coordinates on a \
+    grid plus a small buffer.
+
+    Parameters
+    -----------
+    grid
+        The (y,x) grid of coordinates which determine the Voronoi pixelization's
+    pixelization_grid
+        The (y,x) centre of every Voronoi pixel in scaleds.
+    origin
+        The scaled origin of the Voronoi pixelization's coordinate system.
+    pixel_neighbors
+        An array of length (voronoi_pixels) which provides the index of all neighbors of every pixel in \
+        the Voronoi grid (entries of -1 correspond to no neighbor).
+    pixel_neighbors.sizes
+        An array of length (voronoi_pixels) which gives the number of neighbors of every pixel in the \
+        Voronoi grid.
+    """
+
     @cached_property
     def voronoi(self) -> scipy.spatial.Voronoi:
         """
@@ -337,7 +339,7 @@ class Grid2DVoronoi(AbstractStructure2D):
         return PixelNeighbors(arr=neighbors.astype("int"), sizes=sizes.astype("int"))
 
     @cached_property
-    def splitted_pixelization_grid(self):
+    def split_cross(self):
         half_region_area_sqrt_lengths = 0.5 * np.sqrt(self.pixel_areas)
         # "0.5" is arbitrarily chosen here.
 
@@ -447,7 +449,12 @@ class Grid2DVoronoi(AbstractStructure2D):
         )
 
 
-class Grid2DVoronoiNN(AbstractStructure2D):
+class Grid2DVoronoiNN(Grid2DVoronoi):
+
+    pass
+
+
+class Grid2DDelaunay(AbstractGrid2DIrregular):
     """
     Returns the geometry of the Voronoi pixelization, by alligning it with the outer-most coordinates on a \
     grid plus a small buffer.
@@ -467,263 +474,6 @@ class Grid2DVoronoiNN(AbstractStructure2D):
         An array of length (voronoi_pixels) which gives the number of neighbors of every pixel in the \
         Voronoi grid.
     """
-
-    def __new__(
-        cls,
-        grid: Union[np.ndarray, List],
-        nearest_pixelization_index_for_slim_index: Optional[np.ndarray] = None,
-        *args,
-        **kwargs
-    ):
-        """
-        A grid of (y,x) coordinates which represent a Voronoi pixelization.
-
-        A `Grid2DVoronoi` is ordered arbitrarily, given that there is no regular pattern for a Voronoi mesh's indexing
-        to follow.
-
-        This class is used in conjuction with the `inversion/pixelizations` package to create Voronoi pixelizations
-        and mappers that perform an `Inversion`.
-
-        Parameters
-        -----------
-        grid
-            The grid of (y,x) coordinates corresponding to the centres of each pixel in the Voronoi pixelization.
-        nearest_pixelization_index_for_slim_index
-            When a Voronoi grid is used to create a mapper and inversion, there are mappings between the `data` pixels
-            and Voronoi pixelization. This array contains these mappings and it is used to speed up the creation of the
-            mapper.
-        """
-
-        if type(grid) is list:
-            grid = np.asarray(grid)
-
-        obj = grid.view(cls)
-        obj.nearest_pixelization_index_for_slim_index = (
-            nearest_pixelization_index_for_slim_index
-        )
-
-        return obj
-
-    def __array_finalize__(self, obj: object):
-        """
-        Ensures that the attribute `nearest_pixelization_index_for_slim_index` which is used to speed up the
-        computation of a mapper from the Voronoi grid is not lost in various calculations.
-        """
-        if hasattr(obj, "nearest_pixelization_index_for_slim_index"):
-            self.nearest_pixelization_index_for_slim_index = (
-                obj.nearest_pixelization_index_for_slim_index
-            )
-
-    @cached_property
-    def voronoi(self) -> scipy.spatial.Voronoi:
-        """
-        Returns a `scipy.spatial.Voronoi` object from the (y,x) grid of coordinates which correspond to the centre
-        of every cell of the Voronoi pixelization.
-
-        This object contains numerous attributes describing a Voronoi pixelization, however only the `ridge_points`
-        attribute is used by PyAutoArray (see the Scipy documentation for a description of this attribute).
-
-        There are numerous exceptions that `scipy.spatial.Voronoi` may raise when the input grid of coordinates used
-        to compute the Voronoi pixelization are ill posed. These exceptions are caught and combined into a single
-        `PixelizationException`, which helps exception handling in the `inversion` package.
-        """
-        try:
-            return scipy.spatial.Voronoi(
-                np.asarray([self[:, 1], self[:, 0]]).T, qhull_options="Qbb Qc Qx Qm"
-            )
-        except (ValueError, OverflowError, scipy.spatial.qhull.QhullError) as e:
-            raise exc.PixelizationException() from e
-
-    @cached_property
-    def pixel_neighbors(self) -> PixelNeighbors:
-        """
-        A class packing the ndarrays describing the neighbors of every pixel in the Voronoi pixelization (see
-        `PixelNeighbors` for a complete description of the neighboring scheme).
-
-        The neighbors of a Voronoi pixelization are using the `ridge_points` attribute of the scipy `Voronoi` object,
-        as described in the method `pixelization_util.voronoi_neighbors_from`.
-        """
-        neighbors, sizes = pixelization_util.voronoi_neighbors_from(
-            pixels=self.pixels, ridge_points=np.asarray(self.voronoi.ridge_points)
-        )
-
-        return PixelNeighbors(arr=neighbors.astype("int"), sizes=sizes.astype("int"))
-
-    @cached_property
-    def splitted_pixelization_grid(self):
-        half_region_area_sqrt_lengths = 0.5 * np.sqrt(self.pixel_areas)
-        # "0.5" is arbitrarily chosen here.
-
-        splitted_array = np.zeros((self.pixels, 4, 2))
-
-        splitted_array[:, 0][:, 0] = self[:, 0] + half_region_area_sqrt_lengths
-        splitted_array[:, 0][:, 1] = self[:, 1]
-
-        splitted_array[:, 1][:, 0] = self[:, 0] - half_region_area_sqrt_lengths
-        splitted_array[:, 1][:, 1] = self[:, 1]
-
-        splitted_array[:, 2][:, 0] = self[:, 0]
-        splitted_array[:, 2][:, 1] = self[:, 1] + half_region_area_sqrt_lengths
-
-        splitted_array[:, 3][:, 0] = self[:, 0]
-        splitted_array[:, 3][:, 1] = self[:, 1] - half_region_area_sqrt_lengths
-
-        # print(splitted_array)
-
-        return splitted_array.reshape((self.pixels * 4, 2))
-
-    @cached_property
-    def pixel_areas(self):
-        voronoi_vertices = self.voronoi.vertices
-        voronoi_regions = self.voronoi.regions
-        voronoi_point_region = self.voronoi.point_region
-        region_areas = np.zeros(self.pixels)
-
-        for i in range(self.pixels):
-            region_vertices_indexes = voronoi_regions[voronoi_point_region[i]]
-            if -1 in region_vertices_indexes:
-                region_areas[i] = -1
-            else:
-                region_areas[i] = grid_2d_util.compute_polygon_area(
-                    voronoi_vertices[region_vertices_indexes]
-                )
-
-        max_area = np.percentile(region_areas, 90.0)
-        # We set the maximum area to be 90.0 percentile of the areas. Sometimes, pixels at boundaries can have very large unrealistic areas.
-
-        region_areas[region_areas == -1] = max_area
-        region_areas[region_areas > max_area] = max_area
-
-        return region_areas
-
-    @property
-    def origin(self) -> Tuple[float, float]:
-        """
-        The (y,x) origin of the Voronoi grid, which is fixed to (0.0, 0.0) for simplicity.
-        """
-        return 0.0, 0.0
-
-    @property
-    def pixels(self) -> int:
-        """
-        The total number of pixels in the Voronoi pixelization.
-        """
-        return self.shape[0]
-
-    @classmethod
-    def manual_slim(cls, grid) -> "Grid2DVoronoi":
-        """
-        Convenience method which mimicks the API of other `Grid2D` objects in PyAutoArray.
-        """
-        return Grid2DVoronoi(grid=grid)
-
-    @property
-    def shape_native_scaled(self) -> Tuple[float, float]:
-        """
-        The (y,x) 2D shape of the Voronoi pixelization in scaled units, computed from the minimum and maximum y and x v
-        alues of the pixelization.
-        """
-        return (
-            np.amax(self[:, 0]).astype("float") - np.amin(self[:, 0]).astype("float"),
-            np.amax(self[:, 1]).astype("float") - np.amin(self[:, 1]).astype("float"),
-        )
-
-    @property
-    def scaled_maxima(self) -> Tuple[float, float]:
-        """
-        The maximum (y,x) values of the Voronoi pixelization in scaled coordinates returned as a tuple (y_max, x_max).
-        """
-        return (
-            np.amax(self[:, 0]).astype("float"),
-            np.amax(self[:, 1]).astype("float"),
-        )
-
-    @property
-    def scaled_minima(self) -> Tuple[float, float]:
-        """
-        The minimum (y,x) values of the Voronoi pixelization in scaled coordinates returned as a tuple (y_min, x_min).
-        """
-        return (
-            np.amin(self[:, 0]).astype("float"),
-            np.amin(self[:, 1]).astype("float"),
-        )
-
-    @property
-    def extent(self) -> np.ndarray:
-        return np.array(
-            [
-                self.scaled_minima[1],
-                self.scaled_maxima[1],
-                self.scaled_minima[0],
-                self.scaled_maxima[0],
-            ]
-        )
-
-
-class Grid2DDelaunay(AbstractStructure2D):
-    """
-    Returns the geometry of the Voronoi pixelization, by alligning it with the outer-most coordinates on a \
-    grid plus a small buffer.
-
-    Parameters
-    -----------
-    grid
-        The (y,x) grid of coordinates which determine the Voronoi pixelization's
-    pixelization_grid
-        The (y,x) centre of every Voronoi pixel in scaleds.
-    origin
-        The scaled origin of the Voronoi pixelization's coordinate system.
-    pixel_neighbors
-        An array of length (voronoi_pixels) which provides the index of all neighbors of every pixel in \
-        the Voronoi grid (entries of -1 correspond to no neighbor).
-    pixel_neighbors.sizes
-        An array of length (voronoi_pixels) which gives the number of neighbors of every pixel in the \
-        Voronoi grid.
-    """
-
-    def __new__(
-        cls,
-        grid: Union[np.ndarray, List],
-        nearest_pixelization_index_for_slim_index: Optional[np.ndarray] = None,
-        *args,
-        **kwargs
-    ):
-        """
-        A pixelization-grid of (y,x) coordinates which are used to form the pixel centres of adaptive pixelizations in the \
-        *pixelizations* module.
-
-        A `Grid2DRectangular` is ordered such pixels begin from the top-row of the mask and go rightwards and then \
-        downwards. Therefore, it is a ndarray of shape [total_pix_pixels, 2]. The first element of the ndarray \
-        thus corresponds to the pixelization pixel index and second element the y or x arc -econd coordinates. For example:
-
-        - pix_grid[3,0] = the 4th unmasked pixel's y-coordinate.
-        - pix_grid[6,1] = the 7th unmasked pixel's x-coordinate.
-
-        Parameters
-        -----------
-        pix_grid
-            The grid of (y,x) scaled coordinates of every image-plane pixelization grid used for adaptive source \
-            -plane pixelizations.
-        nearest_pixelization_index_for_slim_index
-            A 1D array that maps every grid pixel to its nearest pixelization-grid pixel.
-        """
-
-        if type(grid) is list:
-            grid = np.asarray(grid)
-
-        obj = grid.view(cls)
-        obj.nearest_pixelization_index_for_slim_index = (
-            nearest_pixelization_index_for_slim_index
-        )
-
-        return obj
-
-    def __array_finalize__(self, obj: object):
-
-        if hasattr(obj, "nearest_pixelization_index_for_slim_index"):
-            self.nearest_pixelization_index_for_slim_index = (
-                obj.nearest_pixelization_index_for_slim_index
-            )
 
     @cached_property
     def Delaunay(self) -> scipy.spatial.Delaunay:
@@ -755,7 +505,7 @@ class Grid2DDelaunay(AbstractStructure2D):
         return PixelNeighbors(arr=neighbors.astype("int"), sizes=sizes.astype("int"))
 
     @cached_property
-    def splitted_pixelization_grid(self):
+    def split_cross(self):
         half_region_area_sqrt_lengths = 0.5 * np.sqrt(self.pixel_areas)
         # "0.5" is arbitrarily chosen here.
